+++ modules/xmcp/gitIntegration/mdmimpl/WorkspaceObjectManagementImpl/src/xmcp/gitintegration/impl/WorkspaceObjectManagementServiceOperationImpl.java
@@ -39,9 +39,9 @@ import xmcp.gitintegration.WorkspaceObjectManagementServiceOperation;
 import xmcp.gitintegration.WorkspaceXmlCreationConfig;
 import xmcp.gitintegration.WorkspaceXmlPath;
 import xmcp.gitintegration.cli.generated.OverallInformationProvider;
-import xmcp.gitintegration.cli.tools.CreateWorkspaceXmlTools;
-import xmcp.gitintegration.cli.tools.ResolveWorkspaceDiffsTools;
-import xmcp.gitintegration.cli.tools.WorkspaceStatusTools;
+import xmcp.gitintegration.tools.CreateWorkspaceXmlTools;
+import xmcp.gitintegration.tools.ResolveWorkspaceDiffsTools;
+import xmcp.gitintegration.tools.WorkspaceStatusTools;
 import xmcp.gitintegration.impl.processing.WorkspaceContentProcessingPortal;
 import xmcp.gitintegration.repository.RepositoryConnection;
 import xmcp.gitintegration.storage.WorkspaceDifferenceListStorage;
